feat: implement Blog and Calendar components for improved content display
