+++ src/routes/+page.svelte
@@ -1,28 +1,14 @@
 <script lang="ts">
 	import { je_cms_base_url, name, pronouns } from '$lib/store.js';
-	import CalendarListSkeleton from '$lib/components/skeleton/CalendarListSkeleton.svelte';
-	import BlogCardSkeleton from '$lib/components/skeleton/BlogCardSkeleton.svelte';
-	import Slogan from '$lib/components/blocks/Slogan.svelte';
-	import Image from '$lib/components/Image.svelte';
 	import { FormatDate } from '$lib/util/date.js';
+	import Calendar from '$lib/components/blocks/Calendar.svelte';
+	import Blog from '$lib/components/blocks/Blog.svelte';
+	import Image from '$lib/components/Image.svelte';
+	import Slogan from '$lib/components/blocks/Slogan.svelte';
 	import SocialIcons from '$lib/components/blocks/SocialIcons.svelte';
-	import Icon from '$lib/components/icons/Icon.svelte';
 
 	export let data;
 	const { posts, events } = data;
-
-	// Helper to get cover image URL
-	function coverUrl(item: any): string | null {
-		const cover = item?.cover;
-		if (!cover) return null;
-
-		const rel = cover.formats?.small?.url ?? cover.formats?.thumbnail?.url ?? cover.url ?? null;
-
-		if (!rel) return null;
-		if (rel.startsWith('http://') || rel.startsWith('https://')) return rel;
-
-		return `${je_cms_base_url}${rel.startsWith('/') ? '' : '/'}${rel}`;
-	}
 
 	// Helper to check if two ISO date strings are on the same day
 	function sameDay(aIso?: string | null, bIso?: string | null): boolean {
@@ -61,41 +47,6 @@
 		const start = new Date(startIso);
 		const end = endIso ? new Date(endIso) : new Date(start.getTime() + 90 * 60 * 1000);
 		return now >= start && now <= end;
-	}
-
-	// Helper to get label and class for event state
-	function stateLabel(state?: string | null): {
-		label: string;
-		class: string;
-		icon: 'meet' | 'recommend' | 'cancelled' | 'info';
-	} {
-		const base = 'inline-flex items-center gap-1.5 text-xs px-2 py-1 rounded-full border';
-		switch (state) {
-			case 'confirmed':
-				return {
-					label: 'Hier kannst du mich treffen',
-					icon: 'meet',
-					class: `${base} bg-emerald-50 text-emerald-700 border-emerald-200 dark:bg-emerald-950/40 dark:text-emerald-200 dark:border-emerald-900/60`
-				};
-			case 'tentative':
-				return {
-					label: 'Terminempfehlung',
-					icon: 'recommend',
-					class: `${base} bg-amber-50 text-amber-700 border-amber-200 dark:bg-amber-950/35 dark:text-amber-200 dark:border-amber-900/60`
-				};
-			case 'cancelled':
-				return {
-					label: 'Abgesagt',
-					icon: 'cancelled',
-					class: `${base} bg-rose-50 text-rose-700 border-rose-200 dark:bg-rose-950/35 dark:text-rose-200 dark:border-rose-900/60`
-				};
-			default:
-				return {
-					label: state ?? 'Status offen',
-					icon: 'info',
-					class: `${base} bg-zinc-50 text-zinc-700 border-zinc-200 dark:bg-zinc-900/40 dark:text-zinc-300 dark:border-zinc-800`
-				};
-		}
 	}
 </script>
 
@@ -401,276 +352,11 @@
 
 	<!-- Calendar Preview -->
 	<div class="lg:col-span-6">
-		{#await events}
-			<CalendarListSkeleton items={8} />
-		{:then eventsLoaded}
-			<section class="rounded-2xl border border-zinc-200/70 dark:border-zinc-800/70 p-6">
-				<div class="flex items-center justify-between">
-					<div>
-						<div class="text-sm font-medium">Nächste Termine</div>
-						<div class="mt-1 text-sm text-zinc-600 dark:text-zinc-400">Öffentliche Einträge</div>
-					</div>
-				</div>
-
-				<div class="mt-5 space-y-3">
-					{#each eventsLoaded as event (event.id)}
-						{@const eventState = stateLabel(event.state)}
-						{@const today = isTodayEvent(event.start, event.end)}
-						{@const live = isLiveNow(event.start, event.end)}
-						<a
-							href={'/calendar/' + event.documentId}
-							class={`block rounded-xl border transition overflow-hidden
-								${
-									live
-										? 'border-accent/50 ring-2 ring-accent/30 bg-accent/5 dark:bg-accent/10'
-										: today
-											? 'border-accent/30 bg-accent/0 dark:bg-accent/5'
-											: 'border-zinc-200/70 dark:border-zinc-800/70 hover:bg-zinc-50 dark:hover:bg-zinc-900/40'
-								}`}
-						>
-							<div class="flex items-center gap-4 p-4">
-								<div
-									class="hidden sm:grid rounded-2xl w-16 h-16 border border-zinc-200/70 dark:border-zinc-800/70 bg-zinc-50 dark:bg-zinc-900/30 place-items-center px-2"
-								>
-									<div class="text-center leading-none">
-										<div class="text-lg font-semibold">
-											{FormatDate(event.start, 'day')}
-										</div>
-										<div class="mt-1 text-[11px] text-zinc-600 dark:text-zinc-400">
-											{FormatDate(event.start, 'monthshort')}
-										</div>
-									</div>
-								</div>
-
-								{#if coverUrl(event)}
-									<div>
-										<Image
-											src={coverUrl(event)!}
-											alt={event.cover?.alternativeText ?? event.subject}
-											classNames="h-16 w-16 rounded-2xl object-cover border border-zinc-200/70 dark:border-zinc-800/70"
-											loading="lazy"
-											copyright={event.copyright?.enabled
-												? [
-														{
-															enabled: event.copyright?.enabled,
-															name: event.copyright?.name,
-															url: event.copyright?.url || '',
-															compact: true
-														}
-													]
-												: []}
-										/>
-									</div>
-								{/if}
-
-								<div class="min-w-0 flex-1">
-									<!-- Label: floated rechts -->
-									<span
-										class={`${eventState.class} float-right ml-3 mt-0.5`}
-										title={eventState.label}
-									>
-										{#if eventState.icon === 'meet'}
-											<svg
-												class="h-3.5 w-3.5"
-												viewBox="0 0 24 24"
-												fill="none"
-												stroke="currentColor"
-												stroke-width="2"
-												aria-hidden="true"
-											>
-												<path d="M16 21v-2a4 4 0 0 0-4-4H6a4 4 0 0 0-4 4v2" />
-												<circle cx="9" cy="7" r="4" />
-												<path d="M19 8v6" />
-												<path d="M22 11h-6" />
-											</svg>
-										{:else if eventState.icon === 'recommend'}
-											<svg
-												class="h-3.5 w-3.5"
-												viewBox="0 0 24 24"
-												fill="none"
-												stroke="currentColor"
-												stroke-width="2"
-												aria-hidden="true"
-											>
-												<path
-													d="M12 17.3l-6.2 3.3 1.2-6.9L1.8 9l7-1L12 1.7 15.2 8l7 1-5.2 4.7 1.2 6.9z"
-												/>
-											</svg>
-										{:else if eventState.icon === 'cancelled'}
-											<svg
-												class="h-3.5 w-3.5"
-												viewBox="0 0 24 24"
-												fill="none"
-												stroke="currentColor"
-												stroke-width="2"
-												aria-hidden="true"
-											>
-												<path d="M18 6L6 18" />
-												<path d="M6 6l12 12" />
-											</svg>
-										{:else}
-											<svg
-												class="h-3.5 w-3.5"
-												viewBox="0 0 24 24"
-												fill="none"
-												stroke="currentColor"
-												stroke-width="2"
-												aria-hidden="true"
-											>
-												<circle cx="12" cy="12" r="10" />
-												<path d="M12 16v-4" />
-												<path d="M12 8h.01" />
-											</svg>
-										{/if}
-
-										<span class="max-w-20.5 text-balance hyphens-auto wrap-break-word"
-											>{eventState.label}</span
-										>
-										{#if live}
-											<span
-												class="mr-1 inline-flex items-center gap-1 text-[11px] font-semibold text-accent"
-											>
-												<span class="h-1.5 w-1.5 rounded-full bg-accent animate-pulse"></span>
-												LIVE
-											</span>
-										{/if}
-									</span>
-									{#if live}
-										<span
-											class="float-right ml-2 mt-0.5 inline-flex items-center gap-1 rounded-full px-2 py-1 text-xs bg-accent/10 text-accent border border-accent/30"
-										>
-											<span class="h-1.5 w-1.5 rounded-full bg-accent animate-pulse"></span>
-											Jetzt
-										</span>
-									{/if}
-
-									<!-- Titel fließt links, 2 Zeilen -->
-									<div class="font-medium text-balance hyphens-auto wrap-break-word">
-										{event.subject}
-									</div>
-
-									<!-- Clear float, damit Meta darunter sauber startet -->
-									<div class="clear-both"></div>
-
-									<!-- Datum / Ort -->
-									<div
-										class="mt-1 text-sm text-zinc-600 dark:text-zinc-400 flex flex-wrap gap-x-3 gap-y-1"
-									>
-										<span>
-											{FormatDate(event.start, 'date')}
-											<span class="text-zinc-500 dark:text-zinc-400">
-												· {FormatDate(event.start, 'time')}
-												{#if event.end}
-													{#if sameDay(event.start, event.end)}
-														– {FormatDate(event.end, 'time')}
-													{:else}
-														– {FormatDate(event.end, 'date')} {FormatDate(event.end, 'time')}
-													{/if}
-												{/if}
-											</span>
-										</span>
-
-										{#if event.location}
-											<span class="text-zinc-500 dark:text-zinc-400 truncate"
-												>· {event.location}</span
-											>
-										{/if}
-									</div>
-
-									{#if event.description}
-										<p class="mt-2 text-sm text-zinc-600 dark:text-zinc-400 line-clamp-2">
-											{event.description}
-										</p>
-									{/if}
-								</div>
-							</div></a
-						>
-					{/each}
-				</div>
-			</section>
-		{:catch err}
-			<div class="rounded-2xl border border-zinc-200/70 dark:border-zinc-800/70 p-6">
-				<div class="text-sm font-medium">Kalender konnte nicht geladen werden</div>
-				<div class="mt-1 text-sm text-zinc-600 dark:text-zinc-400">{String(err)}</div>
-			</div>
-		{/await}
+		<Calendar {events} items={8} />
 	</div>
 
 	<!-- Blog Preview -->
 	<div class="lg:col-span-6">
-		<section class="grid gap-5 sm:grid-cols-2">
-			{#await posts}
-				<BlogCardSkeleton items={4} />
-			{:then ps}
-				{#each ps.slice(0, 4) as post (post.id ?? post.documentId)}
-					<a
-						href={`/blog/${post.documentId}`}
-						class="group block rounded-2xl border border-zinc-200/70 dark:border-zinc-800/70 overflow-hidden hover:bg-zinc-50 dark:hover:bg-zinc-900/40 transition focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/40"
-						aria-label={`Blogbeitrag öffnen: ${post.title}`}
-					>
-						{#if coverUrl(post)}
-							<div class="relative">
-								<Image
-									src={coverUrl(post)!}
-									alt={post.cover?.alternativeText ?? post.title ?? 'Blogbild'}
-									classNames="block w-full aspect-video object-cover"
-									loading="lazy"
-									copyright={post.copyright?.enabled
-										? [
-												{
-													enabled: post.copyright.enabled,
-													name: post.copyright.name,
-													url: post.copyright.url || '',
-													compact: true,
-													size: 'xs'
-												}
-											]
-										: []}
-								/>
-							</div>
-						{/if}
-
-						<div class="p-5">
-							<div
-								class="inline-flex items-center gap-2 text-xs px-3 py-1 rounded-full bg-accent/10 text-accent border border-accent/30"
-							>
-								<Icon name={post.type} classes="h-4 w-4" />
-								{#if post.type == 'article'}
-									Artikel
-								{:else if post.type == 'comment'}
-									Kommentar
-								{:else}
-									<span>Beitrag</span>
-								{/if}
-							</div>
-
-							<h2 class="mt-3 text-lg font-semibold tracking-tight line-clamp-2">
-								{post.title}
-							</h2>
-
-							<p class="mt-2 text-sm text-zinc-600 dark:text-zinc-400 line-clamp-2">
-								{post.description}
-							</p>
-
-							<div class="mt-4">
-								<span
-									class="inline-flex items-center gap-1 text-sm font-medium text-accent group-hover:underline"
-								>
-									Weiterlesen
-									<Icon name="arrow-right" classes="w-4 h-4" />
-								</span>
-							</div>
-						</div>
-					</a>
-				{/each}
-			{:catch err}
-				<div
-					class="sm:col-span-2 rounded-2xl border border-zinc-200/70 dark:border-zinc-800/70 p-6"
-				>
-					<div class="text-sm font-medium">Blog konnte nicht geladen werden</div>
-					<div class="mt-1 text-sm text-zinc-600 dark:text-zinc-400">{String(err)}</div>
-				</div>
-			{/await}
-		</section>
+		<Blog {posts} limit={4} />
 	</div>
 </section>
